Fixes to various structures to use strings for the file signature. Fixes using, fixes display bug with SizeSpan.ToString.

+++ src/BinaryMapper.Windows.Executable/Structures/IMAGE_RESOURCE_DIRECTORY.cs
@@ -1,6 +1,5 @@
-﻿using WORD = System.UInt16;
-using DWORD = System.UInt32;
-using BinaryMapper.Core.Attributes;
+﻿using DWORD = System.UInt32;
+using WORD = System.UInt16;
 
 namespace BinaryMapper.Windows.Executable.Structures
 {
